refactor(git-provider): split into git-provider, git-install, git-stash (D4)
Decompose monolithic git.ts (799 LOC, 3 responsibilities) into focused modules:
- git-provider.ts: SourceProvider, repo URL parsing, cache/mirror sync
- git-install.ts: install/clone helpers (akm add/update operations)
- git-stash.ts: stash save/commit/push operations

Barrel export (git.ts) maintains consumer API stability.
